% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEApdata.R
\docType{data}
\name{mset_Reactome_PFA_format_HMDB}
\alias{mset_Reactome_PFA_format_HMDB}
\title{A metabolite set (HMDB IDs) based on Reactome}
\description{
A metabolite set for Plasmodium falciparum based on Reactome project. "Reactome" https://reactome.org/download-data/ (2018 Aug 5).
We converted https://reactome.org/download/current/ChEBI2Reactome.txt to
our metabolite set with HMDB IDs by using
https://github.com/afukushima/MSEApdata/blob/master/data-raw/reactome2cmt.py
This dataset contains 192 metabolite sets.
}
\references{
Fabregat et al. 2018 PMID: 29145629
}
\author{
Kozo Nishida
}
\keyword{datasets}
